Consolidate welcome notification specs into user_spec
Moves the contents of user_send_welcome_notifications_spec.rb into
user_spec.rb under a `#send_welcome_notifications` describe block,
following the standard one-spec-per-model convention.
